feat: add CalendarItemsList component and types
  - Added CalendarItemsList and moved logic of items to the separate
component
  - Added types and styles files
