+++ src/components/index.ts
@@ -1,4 +1,5 @@
 export { CalendarItem } from './CalendarItem';
+export { CalendarItemsList } from './CalendarItemsList';
 export { Header } from './Header';
 export { Input } from './Input';
 export { Modal } from './Modal';
